+++ ui/__init__.py
@@ -18,12 +18,15 @@ from ui.info_pages import screen_about, screen_tutorial
 from ui.upload_database import screen_upload_database
 from ui.main_menus import (
     screen_orthography_hub,
+    screen_collect_lexical_data_hub,
     screen_main_menu_phonological,
     screen_main_menu_orthography,
     screen_main_menu_acoustic,
+    screen_main_menu_collect,
     screen_task_placeholder,
 )
 from ui.create_data_collection_script import screen_create_data_collection_script
+from ui.create_prompts_list import screen_create_prompts_list
 from ui.add_audio_links import screen_add_audio_links
 from ui.update_entries_from_words import screen_update_entries_from_words
 from ui.explode_entries import screen_explode_entries
@@ -40,12 +43,15 @@ SCREEN_FUNCTIONS: dict[str, Callable] = {
     Screen.MAIN_MENU_PHONOLOGICAL: screen_main_menu_phonological,
     Screen.MAIN_MENU_ORTHOGRAPHY: screen_main_menu_orthography,
     Screen.MAIN_MENU_ACOUSTIC: screen_main_menu_acoustic,
+    Screen.MAIN_MENU_COLLECT: screen_main_menu_collect,
     Screen.ORTHOGRAPHY_HUB: screen_orthography_hub,
+    Screen.COLLECT_LEXICAL_DATA_HUB: screen_collect_lexical_data_hub,
     Screen.SCRIPT_CONVERTER: screen_script_converter,
     Screen.ABOUT: screen_about,
     Screen.TUTORIAL: screen_tutorial,
     Screen.TASK_PLACEHOLDER: screen_task_placeholder,
     Screen.TASK_CREATE_DATA_COLLECTION_SCRIPT: screen_create_data_collection_script,
+    Screen.TASK_CREATE_PROMPTS_LIST: screen_create_prompts_list,
     Screen.TASK_ADD_AUDIO_LINKS: screen_add_audio_links,
     Screen.TASK_UPDATE_ENTRIES_FROM_WORDS: screen_update_entries_from_words,
     Screen.TASK_EXPLODE_ENTRIES: screen_explode_entries,